Styra/Opa: Split apart the batch types from OpaClient.
