+++ Styra/Opa/OpaClient.cs
@@ -9,188 +9,6 @@ using Styra.Opa.OpenApi.Models.Errors;
 
 namespace Styra.Opa;
 
-public class OpaResult
-{
-    /// <summary>
-    /// If decision logging is enabled, this field contains a string that uniquely identifies the decision. The identifier will be included in the decision log event for this decision. Callers can use the identifier for correlation purposes.
-    /// </summary>
-    [JsonProperty("decision_id")]
-    public string? DecisionId { get; set; }
-
-    /// <summary>
-    /// If query metrics are enabled, this field contains query performance metrics collected during the parse, compile, and evaluation steps.
-    /// </summary>
-    [JsonProperty("metrics")]
-    public Dictionary<string, object>? Metrics { get; set; }
-
-    /// <summary>
-    /// Provenance information can be requested on individual API calls and are returned inline with the API response. To obtain provenance information on an API call, specify the `provenance=true` query parameter when executing the API call.
-    /// </summary>
-    [JsonProperty("provenance")]
-    public Provenance? Provenance { get; set; }
-
-    /// <summary>
-    /// The base or virtual document referred to by the URL path. If the path is undefined, this key will be omitted.
-    /// </summary>
-    [JsonProperty("result")]
-    public Result? Result { get; set; }
-
-    /// <summary>
-    /// The HTTP status code for the request. Limited to "200" or "500".
-    /// </summary>
-    [JsonProperty("http_status_code")]
-    public string? HttpStatusCode { get; set; }
-
-    public OpaResult() { }
-
-    public OpaResult(ResponsesSuccessfulPolicyResponse resp)
-    {
-        DecisionId = resp.DecisionId;
-        Metrics = resp.Metrics;
-        Provenance = resp.Provenance;
-        Result = resp.Result;
-        HttpStatusCode = resp.HttpStatusCode;
-    }
-
-    public OpaResult(SuccessfulPolicyResponse resp)
-    {
-        DecisionId = resp.DecisionId;
-        Metrics = resp.Metrics;
-        Provenance = resp.Provenance;
-        Result = resp.Result;
-    }
-
-    public static explicit operator OpaResult(ResponsesSuccessfulPolicyResponse e) => new OpaResult(e);
-    public static explicit operator OpaResult(SuccessfulPolicyResponse e) => new OpaResult(e);
-
-    public override string ToString()
-    {
-        return JsonConvert.SerializeObject(this);
-    }
-}
-
-public class OpaError
-{
-    /// <summary>
-    /// The short-form category of error, such as "internal_error", "invalid_policy_or_data", etc.
-    /// </summary>
-    [JsonProperty("code")]
-    public string Code { get; set; } = default!;
-
-    /// <summary>
-    /// If decision logging is enabled, this field contains a string that uniquely identifies the decision. The identifier will be included in the decision log event for this decision. Callers can use the identifier for correlation purposes.
-    /// </summary>
-    [JsonProperty("decision_id")]
-    public string? DecisionId { get; set; }
-
-    /// <summary>
-    /// The long-form error message from the OPA instance, describing what went wrong.
-    /// </summary>
-    [JsonProperty("message")]
-    public string Message { get; set; } = default!;
-
-    /// <summary>
-    /// The HTTP status code for the request. Limited to "200" or "500".
-    /// </summary>
-    [JsonProperty("http_status_code")]
-    public string? HttpStatusCode { get; set; }
-
-    public OpaError()
-    {
-
-    }
-    public OpaError(Styra.Opa.OpenApi.Models.Components.ServerError err)
-    {
-        Code = err.Code;
-        DecisionId = err.DecisionId;
-        Message = err.Message;
-        HttpStatusCode = err.HttpStatusCode;
-    }
-    public OpaError(Styra.Opa.OpenApi.Models.Errors.ServerError err)
-    {
-        Code = err.Code;
-        DecisionId = err.DecisionId;
-        Message = err.Message;
-    }
-
-    public static explicit operator OpaError(OpenApi.Models.Components.ServerError e) => new OpaError(e);
-    public static explicit operator OpaError(OpenApi.Models.Errors.ServerError e) => new OpaError(e);
-
-    public override string ToString()
-    {
-        return JsonConvert.SerializeObject(this);
-    }
-}
-
-public class OpaBatchInputs : Dictionary<string, Dictionary<string, object>>
-{
-    public override string ToString()
-    {
-        return JsonConvert.SerializeObject(this);
-    }
-}
-
-public class OpaBatchResults : Dictionary<string, OpaResult>
-{
-    public override string ToString()
-    {
-        return JsonConvert.SerializeObject(this);
-    }
-}
-
-public class OpaBatchErrors : Dictionary<string, OpaError>
-{
-    public override string ToString()
-    {
-        return JsonConvert.SerializeObject(this);
-    }
-}
-
-// Used for converting inputs for the Batch Query API, and converting result types
-// into useful higher-level types.
-public static class DictionaryExtensions
-{
-    public static Dictionary<string, Input> ToOpaBatchInputRaw(this Dictionary<string, Dictionary<string, object>> inputs)
-    {
-        var opaBatchInputs = new Dictionary<string, Input>();
-        foreach (var kvp in inputs)
-        {
-            opaBatchInputs[kvp.Key] = Input.CreateMapOfAny(kvp.Value);
-        }
-        return opaBatchInputs;
-    }
-
-    public static OpaBatchErrors ToOpaBatchErrors(this Dictionary<string, Styra.Opa.OpenApi.Models.Errors.ServerError> errors)
-    {
-        var opaBatchErrors = new OpaBatchErrors();
-        foreach (var kvp in errors)
-        {
-            opaBatchErrors[kvp.Key] = new OpaError(kvp.Value);
-        }
-        return opaBatchErrors;
-    }
-
-    public static OpaBatchErrors ToOpaBatchErrors(this Dictionary<string, Styra.Opa.OpenApi.Models.Components.ServerError> errors)
-    {
-        var opaBatchErrors = new OpaBatchErrors();
-        foreach (var kvp in errors)
-        {
-            opaBatchErrors[kvp.Key] = new OpaError(kvp.Value);
-        }
-        return opaBatchErrors;
-    }
-
-    public static OpaBatchResults ToOpaBatchResults(this Dictionary<string, SuccessfulPolicyResponse> responses)
-    {
-        var opaBatchResults = new OpaBatchResults();
-        foreach (var kvp in responses)
-        {
-            opaBatchResults[kvp.Key] = (OpaResult)kvp.Value;
-        }
-        return opaBatchResults;
-    }
-}
-
 /// <summary>
 /// OpaClient provides high-level convenience APIs for interacting with an OPA server.
 /// It is generally recommended to use this class for most common OPA integrations.
